// lib/schema.ts
/* eslint-disable @typescript-eslint/consistent-type-definitions */
/* eslint-disable @typescript-eslint/no-redundant-type-constituents */

// ⚠️ This file is auto-generated by Supabase CLI
// Run `npx supabase gen types typescript --project-id <id> > lib/schema.ts`
// DO NOT manually edit unless you are disabling ESLint rules

export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
  public: {
    Tables: {
      profiles: {
        Row: {
          id: string;
          email: string;
          display_name: string;
          biography: string | null;
        };
        Insert: {
          id: string;
          email: string;
          display_name: string;
          biography?: string | null;
        };
        Update: {
          id?: string;
          email?: string;
          display_name?: string;
          biography?: string | null;
        };
      };
      species: {
        Row: {
          id: number;
          scientific_name: string;
          common_name: string | null;
          total_population: number | null;
          kingdom: string;
          description: string | null;
          image: string | null;
          author: string;
        };
        Insert: {
          scientific_name: string;
          common_name?: string | null;
          total_population?: number | null;
          kingdom: string;
          description?: string | null;
          image?: string | null;
          author: string;
        };
        Update: {
          scientific_name?: string;
          common_name?: string | null;
          total_population?: number | null;
          kingdom?: string;
          description?: string | null;
          image?: string | null;
          author?: string;
        };
      };
      comments: {
        Row: {
          id: string;
          species_id: number;
          user_id: string;
          content: string;
          created_at: string;
        };
        Insert: {
          id?: string;
          species_id: number;
          user_id: string;
          content: string;
          created_at?: string;
        };
        Update: {
          id?: string;
          species_id?: number;
          user_id?: string;
          content?: string;
          created_at?: string;
        };
      };
    };
  };
};